test: fix expected result in string reductions
Before this fix, the expected result in certain co_{max,min,reduce}
tests was assumed to be the corresponding global extremum for the
data set. This fix computes the local minimum after truncating the
data set, if necessary, to ensure that the number of elements is
never greater than the number images.