doc(joss): separate Comparison section
The journal's required content includes "A description of how this
software compares to other commonly-used packages in this research
area." This commit separates the recently added comparison into its
own section of the paper.  More content is needed. In particular,
this might be a good place to cite the figure that shows the
software stack.