Fix stop test for flang
F23 standard for `execute_command_line` says optional argument
`cmdstat` is set to:

"a processor-dependent positive value if an error condition occurs,"

and flang's `execute_command_line` apparently fatal errors if the exit code is
non-zero and cmdstat is absent. This was breaking stop subtest
`exit_with_integer_stop_code` which deliberately sends an exit code of 99.

Change our stop test to always pass `cmdstat` so that we can diagnose the
(possible) error ourselves in the test code.